+++ PlayerBaseApi/Models/PlayerBasePlacementDTO.cs
@@ -4,8 +4,8 @@
     {
         public int BuildingLevel { get; set; }
         public string? UpdateEndDate { get; set; } = null;
-        public int CoordX { get; set; }
-        public int CoordY { get; set; }
+        public double CoordX { get; set; }
+        public double CoordY { get; set; }
         public BuildingTypeDTO BuildingType { get; set; }
     }
 }
